Move convert_to_label_encoding to global util.py
This function was used both by dataset generators and by learners.
To be able to make the dataset generation optional, this needs to be decoupled.
