+++ csrank/dataset_reader/discretechoice/letor_listwise_discrete_choice_dataset_reader.py
@@ -3,9 +3,9 @@ import logging
 from sklearn.utils import check_random_state
 
 from csrank.constants import DISCRETE_CHOICE
-from .util import convert_to_label_encoding
 from .util import sub_sampling_discrete_choices_from_relevance
 from ..letor_listwise_dataset_reader import LetorListwiseDatasetReader
+from ...util import convert_to_label_encoding
 
 logger = logging.getLogger(__name__)
