+++ csrank/dataset_reader/discretechoice/sushi_discrete_choice_dataset_reader.py
@@ -4,9 +4,9 @@ from sklearn.model_selection import StratifiedKFold
 from sklearn.utils import check_random_state
 
 from csrank.constants import DISCRETE_CHOICE
-from csrank.dataset_reader.discretechoice.util import convert_to_label_encoding
 from csrank.dataset_reader.util import standardize_features
 from ..sushi_dataset_reader import SushiDatasetReader
+from ...util import convert_to_label_encoding
 
 
 class SushiDiscreteChoiceDatasetReader(SushiDatasetReader):
