+++ csrank/discretechoice/discrete_choice.py
@@ -1,9 +1,9 @@
 from abc import ABCMeta
 
 from csrank.constants import DISCRETE_CHOICE
-from csrank.dataset_reader.discretechoice.util import convert_to_label_encoding
 from csrank.discrete_choice_losses import CategoricalHingeLossMax
 from csrank.learner import SkorchInstanceEstimator
+from csrank.util import convert_to_label_encoding
 
 __all__ = ["DiscreteObjectChooser", "SkorchDiscreteChoiceFunction"]
